+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # Changelog:
 
-* Moved Codefresh ci pipeline into the repo
+* Added cluster commands